requirements.txt
asn1crypto==0.24.0
asyncio==3.4.3
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
Django==2.2
django-cleanup==3.2.0
django-cors-headers==2.5.3
django-rest-framework==0.1.0
django-rest-knox==4.0.1
djangorestframework==3.9.3
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.3
idna==2.8
oauth2client==4.1.3
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
python-magic==0.4.15
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
